# docs/oas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QDash API",
    "summary": "QDash API",
    "description": "API for QDash",
    "contact": {
      "name": "QDash",
      "email": "oqtopus-team@googlegroups.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "List all tasks",
        "description": "List all tasks.\n\nArgs:\n----\n    ctx (ProjectContext): The project context with user and project information.\n    backend (Optional[str]): Optional backend name to filter tasks by.\n\nReturns:\n-------\n    list[TaskResponse]: The list of tasks.",
        "operationId": "listTasks",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "backend",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional backend name to filter tasks by",
              "title": "Backend"
            },
            "description": "Optional backend name to filter tasks by"
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/result": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "Get task result by task ID",
        "description": "Get task result by task_id.\n\nArgs:\n----\n    task_id: The task ID to search for.\n    ctx: The project context with user and project information.\n\nReturns:\n-------\n    TaskResultResponse: The task result information including figure paths.",
        "operationId": "getTaskResult",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InputParameterModel": {
        "properties": {
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "value_type": {
            "type": "string",
            "title": "Value Type",
            "default": "float"
          },
          "value": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "title": "InputParameterModel",
        "description": "Input parameter class."
      },
...
      "ListTaskResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "ListTaskResponse",
        "description": "Response model for a list of tasks."
      },
...
      "TaskResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "input_parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/InputParameterModel"
            },
            "type": "object",
            "title": "Input Parameters"
          },
          "output_parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/InputParameterModel"
            },
            "type": "object",
            "title": "Output Parameters"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "task_type",
          "input_parameters",
          "output_parameters"
        ],
        "title": "TaskResponse",
        "description": "Response model for a task."
      },
...
      "TaskResultResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "task_name": {
            "type": "string",
            "title": "Task Name"
          },
          "qid": {
            "type": "string",
            "title": "Qid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "figure_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Figure Path"
          },
          "json_figure_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Json Figure Path"
          },
          "input_parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Parameters"
          },
          "output_parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Parameters"
          },
          "start_at": {
            "type": "string",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "title": "End At"
          },
          "elapsed_time": {
            "type": "string",
            "title": "Elapsed Time"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "task_name",
          "qid",
          "status",
          "execution_id",
          "figure_path",
          "json_figure_path",
          "input_parameters",
          "output_parameters",
          "start_at",
          "end_at",
          "elapsed_time"
        ],
        "title": "TaskResultResponse",
        "description": "Response model for task result by task_id.\n\nAttributes\n----------\n    task_id (str): The task ID.\n    task_name (str): The name of the task.\n    qid (str): The qubit or coupling ID.\n    status (str): The task status.\n    execution_id (str): The execution ID.\n    figure_path (list[str]): List of figure paths.\n    json_figure_path (list[str]): List of JSON figure paths.\n    input_parameters (dict): Input parameters.\n    output_parameters (dict): Output parameters.\n    start_at (str): Start time.\n    end_at (str): End time.\n    elapsed_time (str): Elapsed time."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}